+++ source/common/upstream/cluster_manager_impl.h
@@ -21,7 +21,7 @@
 #include "envoy/upstream/cluster_manager.h"
 
 #include "common/common/cleanup.h"
-#include "common/config/new_grpc_mux_impl.h"
+#include "common/config/grpc_mux_impl.h"
 #include "common/config/subscription_factory_impl.h"
 #include "common/http/async_client_impl.h"
 #include "common/upstream/load_stats_reporter.h"
